+++ src/librecast.c
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: GPL-2.0-only OR GPL-3.0-only */
-/* Copyright (c) 2017-2021 Brett Sheffield <bacs@librecast.net> */
+/* Copyright (c) 2017-2022 Brett Sheffield <bacs@librecast.net> */
 
 #define _GNU_SOURCE
 #include "librecast_pvt.h"
@@ -9,7 +9,6 @@
 #include <assert.h>
 #include <fcntl.h>
 #include <ifaddrs.h>
-#include <net/if.h>
 #include <netinet/in.h>
 #include <pthread.h>
 #include <signal.h>